Rename `tscript-lang` to `tscript-lib`
