+++ src/lang/tscript-lib/core.ts
