+++ src/lang/tscript-lib/lib-audio.ts
