+++ src/lang/tscript-lib/lib-audio.tscript.ts
