+++ src/lang/tscript-lib/lib-canvas.ts
