+++ src/lang/tscript-lib/lib-canvas.tscript.ts
