+++ src/lang/tscript-lib/lib-math.ts
