+++ src/lang/tscript-lib/lib-math.tscript.ts
